kill errors

+++ database/migrations/2023_08_06_081644_create_entries_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('entries', function (Blueprint $table) {
             $table->id();
             $table->text('content');
-            $table->unsignedSmallInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->timestamps();
 
             $table->foreign('user_id')->on('users')->references('id');
